Add HEAPF32 to Emscripten exported methods to fix raylib audio
Without this export, we see this error when we try to play a sound:

Aborted('HEAPF32' was not exported. add it to EXPORTED_RUNTIME_METHODS (see the Emscripten FAQ)) index.js:697:6
Uncaught RuntimeError: Aborted('HEAPF32' was not exported. add it to EXPORTED_RUNTIME_METHODS (see the Emscripten FAQ))
    abort http://localhost:8000/index.js:715
    get http://localhost:8000/index.js:584
    onaudioprocess http://localhost:8000/index.js:9622
index.js:715:11